Centered footer. Removed redundant images
The footer is now centered - Was previously left aligned to make room for site-specific links. The site does not currently contain these so the placement was out-of-place

The site now displays 6 photos on the Home page instead of 9 - Removed the image that I was unable to scale to 1:1 and the two images found on the AboutUs page.